Update app v2.py
The code correctly imports the time module for time-related operations but it doesn't need to import the ctypes library since it's not used and Instead of opening the hosts file in append mode ("a"), we should open it in write mode ("w") to replace the existing content. This ensures that the file is correctly updated with the blocked websites.  The code uses the current_time.tm_hour attribute to check the current hour.  but  this attribute represents the hour in 24-hour format. To check if the current hour is between 9 AM and 5 PM , we need to modify the condition to current_time.tm_hour >= 9 and current_time.tm_hour < 18.